down interface so setting monitor mode always succeeds
if the interface is in normal, i.e. "managed" mode and up,
setting the monitor mode will fail.
we must first bring the interface down, set monitor mode, and bring
it up again.

TODO: maybe it's possible to check whether the interface is already
in monitor mode, so we can leave everything as is.