restore original mode of wireless interface
if the interface is already in monitor mode and up, we don't have to
do anything. if it's in another mode, we restore the mode after
quitting, and restore the original up/down state.

in case the interface was configured with an ip address/netmask:
those settings are currently not restored.
however that case is pretty unlikely and if so the user supposedly has
wpa_supplicant running, which would interfere with our use anyhow.