fix min/max range
there were almost always cases when after a few packets "max"
(max dbm value) got hugely increased beyond any possible value,
so all real measurements appeared very weak.
this was due to (what wireshark labeled as such) "malformed packets".
these packets had in their radiotap header flags field the following
bits set:
IEEE80211_RADIOTAP_FLAGS (1 byte to read)
IEEE80211_RADIOTAP_CHANNEL (4 bytes to read)
and others later on.
this caused get_dbm_off() to return 5 (1+4 from above), reading the
dbm value from the 5th byte instead of the 6th byte were it was
in fact (and usually) stored.
so we read the value "4" instead of something like "-50".

the radiotap header of the malformed packet follows:
000015002a48080000006c098004b5010000070001
of which the it_present field contains 2a480800.
according to my interpretation the dbm should then be saved in
000015002a48080000006c098004b5010000070001
                          ^^
but it is in the following byte.
unfortunately wireshark doesn't tell what exactly in the packet
is malformed so further information is missing.

the workaround is to process the dbm info only from beacon packets
which apparently always are correct.