fix get_flag_off() to respect alignment of the known radiotap data fields
when seen from the start of the radiotap header, some fields require a
specific alignment. for example the TSFT field, which is almost always
present, enforces that it starts at an offset divisible by 8.
so when a single EXT bit is set (and therefore one additional 32bit
bitmask field, equivalent to it_present) is read, the current offset is
12, so if the TSFT bit is set we need to skip its 8 bytes size starting
from offset 16.

while at it, we also added the values for size and alignment for recent
additions to the radiotap "spec".