fix high CPU load by using blocking pcap IO from a thread
after quite some effort debugging the high CPU load of about 40%
when running even the relatively CPU-friendly ncurses backend,
and trying different audio backends, it turned out that the culprit
was pcap's nonblocking mode. it is apparently implemented in a
very inefficient way, doing very frequent polls with a tiny sleep
(or even no sleep at all).
so the only way to get around that is for us to either poll ourselves
on the fd behind the pcap handle (e.g. using select()), or by simply
putting the process_frame logic into a thread using a blocking IO.

(we needed nonblocking IO because we need to query both the GUI and
pcap from the main loop, and by using blocking pcap IO, the gui could
become irresponsible if there was a low packet count on the monitor
interface).