initial version of netgui
if environment var NETGUI is set, a server socket will be spawned on port
9876. clients can then issue a number of commands:

LIST - get list of networks in json that changed since last LIST command
SELECT bssid - select bssid matching submitted one. select the bssid as if
  one would have used cursor up and down in macgeiger to select the net and
  hit enter.
UNSELECT - make macgeiger switch back to the network overview.
QUIT - quit macgeiger.

the plan is to use this to share data with a webserver, which can render
a html page so macgeiger can be controlled from a mobile phone.