handle frame checksum padded to end of packet
using a ATH9K_HTC device which added the FCS, the process_tags code
got confused since it interpreted the checksum following as additional
tag data. now if the driver signals that FCS is appended, we just shrink
the packet len so the checksum is ignored.

for optimal conditions, we should probably deal with the bad checksum
bit AND deal with the case that the checksum doesn't match the packet
contents (needs CRC32 comparison). the relevant code is in reaver's
check_fcs() in 80211.c.