improve beeping precision by using the selected wlan's scale
min/max for the beeping range will be determined by the best values found
for the network, rather than all measured values. this makes it possible to
find the best antenna orientation even for networks that are in a generally
low dbm area. just like the old method, this relies on the antenna being moved
a little to collect realistic min/max values.

the old behaviour can be toggled with the right cursor key (this might be
subject to change).