preliminary support for displaying AP clients
similar to airodump-ng. currently only mac address is displayed, but
we gather more information than that.
it's imaginable that in the future a client can be selected just
like an AP to find its location/optimum signal (e.g. for a deassoc
attack), but it's already helpful to quickly see whether an AP has
clients without having to fiddle with aircrack commands in another
terminal.